V-02 рефактор структуры
