+++ src/components/common/Footer/Footer.module.scss
@@ -1,0 +1,54 @@
+@use '/src/sass/tools' as *;
+
+.footer {
+
+  @include media($md) {
+    background-color: $black-blue;
+  }
+
+  span {
+    color: $gray;
+    font-size: 0.8rem;
+    line-height: 15px;
+
+    @include media($md) {
+      margin-top: 8px;
+      order: 1;
+    }
+
+    @include media($md) {
+      color: $gray-light;
+    }
+  }
+
+  a {
+    font-size: 0.815rem;
+    line-height: 15px;
+    text-decoration: none;
+    color: $black;
+    @include transition(0.2s,0s);
+    &:hover {
+      color: $green;
+      @include transition(0.2s,0s);
+    }
+
+    @include media($md) {
+      color: $green;
+    }
+  }
+}
+
+.body {
+  padding: 32px 0;
+  @include flex(space-between,center,_);
+
+  @include media($xl) {
+    padding: 32px 4px 38px 0;
+  }
+
+  @include media($md) {
+    padding: 16px 0;
+    @include flex(_,flex-end,column);
+    font-size: 0.75rem;
+  }
+}
