+++ src/components/common/Header/Header.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import CustomLink from '../Link/CustomLink';
+import styles from './Header.module.scss';
+import {ReactComponent as GeoIcon} from '../../../content/svg/geo.svg';
+import Container from '../Container/Container';
+
+function Header() {
+  return (
+    <header className={styles.header}>
+      <Container>
+        <div className={styles.body}>
+          <h1>Need for drive</h1>
+          <CustomLink className={styles.geo}>
+            <>
+              <GeoIcon className="logo" />
+              <span>Ульяновск</span>
+            </>
+          </CustomLink>
+        </div>
+      </Container>
+    </header>
+  );
+}
+
+export default Header;
