+++ src/components/common/NavBar/LangSwitcher/LangSwitcher.module.scss
