--- src/components/elements/common/Link/CustomLink.module.scss
@@ -1,10 +1,0 @@
-@use '../../../../sass/tools' as *;
-
-.link {
-  text-decoration: none;
-  @include transition(0.2s,0s);
-  &:hover {
-    color: $green;
-    @include transition(0.2s,0s);
-  }
-}
