--- src/components/elements/menu/NavBar/NavBar.module.scss
@@ -1,131 +1,0 @@
-@use '../../../../sass/tools' as *;
-
-.switcher {
-  margin-bottom: 16px;
-
-  @include media($xl) {
-    margin-bottom: 22px;
-  }
-}
-
-.nav {
-  @include flex(space-between,center,column);
-  width: 64px;
-  color: $white;
-  background: $shadow-black;
-  z-index: 999;
-
-  @include media($lg) {
-    width: 86px;
-  }
-
-  @include media($md) {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 32px;
-    background-color: transparent;
-  }
-}
-
-.burger {
-  margin-top: 32px;
-
-  @include media($md) {
-    margin: 16px 0px 0px 32px;
-  }
-}
-
-.menu {
-  position: absolute;
-  top: 0;
-  left: 0;
-  @include flex(space-between,_,_);
-  width: 100vw;
-  height: 100vh;
-  transform: translateX(-100%);
-  overflow: auto;
-
-  @include transition(0.5s,0s);
-
-  &.active {
-    transform: translateX(0);
-    @include transition(0.5s,0s);
-  }
-
-  @include media($md) {
-    @include flex(space-between,stretch,column);
-  }
-
-  .switcher {
-    display: none;
-    @include media($md) {
-      position: absolute;
-      top: calc(100% - 10px);
-      left: 10px;
-      transform: translate(0%, -100%);
-      display: inherit;
-      background: $shadow-black;
-    }
-  }
-}
-
-.body {
-  flex: 1 1 auto;
-  padding: 35vh 0px 0px 128px;
-  max-height: 100%;
-  background: $shadow-black;
-
-  @include media($xl) {
-    flex: 1 1 100%;
-    padding: 35vh 0px 0px 97px;
-  }
-
-  @include media($md) {
-    flex: 1 1 100%;
-    padding: 15vh 0px 0px 28px;
-  }
-}
-
-.glass {
-  flex: 0 0 calc(50% - 33px);
-  height: 100%;
-  background: $black-blue;
-  opacity: 0.81;
-
-  @include media($xl) {
-    flex: 0 0 calc(40% - 27px);
-  }
-
-  @include media($lg) {
-    flex: 0 0 0;
-  }
-}
-
-.list {
-  margin-bottom: 38px;
-  li {
-    margin-bottom: 22px;
-  }
-}
-
-.link {
-  font-size: 2rem;
-  font-weight: 500;
-  color: $white;
-
-  @include media($lg) {
-    font-size: 1.75rem;
-  }
-
-  @include media($md) {
-    font-size: 1.4rem;
-  }
-}
-
-.social {
-  @include flex(_,_,_);
-  svg {
-    margin-right: 24px;
-  }
-}
