+++ src/components/pages/Homepage/Homepage.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
-import Slider from '../../elements/homepage/slider-components/Slider/Slider';
+import Slider from './components/slider-components/Slider/Slider';
 import styles from './Homepage.module.scss';
-import Start from '../../elements/homepage/Start/Start';
+import Start from './components/Start/Start';
 import slides from './slides';
 
 function Homepage() {
   return (
-    <main className={styles.homepage}>
+    <div className={styles.homepage}>
       <Start />
       <Slider slides={slides} />
-    </main>
+    </div>
   );
 }
