+++ src/components/pages/Homepage/components/Start/Start.module.scss
@@ -1,0 +1,6 @@
+@use '/src/sass/tools' as *;
+
+.start {
+  @include flex(space-between,_,column);
+  flex: 1 1 auto;
+}
