+++ src/components/pages/Homepage/components/slider-components/Dots/Dots.module.scss
