+++ src/components/pages/Homepage/components/slider-components/Slide/Slide.module.scss
