+++ src/components/pages/Homepage/components/slider-components/Slider/Slider.module.scss
