--- src/components/pages/Layout/Layout.module.scss
@@ -1,7 +1,0 @@
-@use '/src/sass/tools' as *;
-
-.layout {
-  @include flex (_,_,_);
-  min-height: 100%;
-  flex: 1 1 auto;
-}
