--- src/components/pages/Layout/Layout.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import {Outlet} from 'react-router-dom';
-import NavBar from '../../elements/menu/NavBar/NavBar';
-import styles from './Layout.module.scss';
-
-function Layout() {
-  return (
-    <div className={styles.layout}>
-      <NavBar />
-      <Outlet />
-    </div>
-  );
-}
-
-export default Layout;
