+++ src/components/pages/Order/Order.module.scss
@@ -1,0 +1,5 @@
+@use '/src/sass/tools' as *;
+
+.order {
+  flex: 1 1 auto;
+}
