+++ src/components/pages/Order/Order.tsx
@@ -1,10 +1,17 @@
 import React from 'react';
+import OrderBody from './components/OrderBody/OrderBody';
+import Header from '../../common/Header/Header';
+import Container from '../../common/Container/Container';
+import styles from './Order.module.scss';
 
 function Order() {
   return (
-    <main>
-      Здесь должен оформляться заказ!
-    </main>
+    <div className={styles.order}>
+      <Container>
+        <Header />
+        <OrderBody />
+      </Container>
+    </div>
   );
 }
