+++ src/components/pages/Order/components/OrderBody/OrderBody.module.scss
@@ -1,0 +1,5 @@
+@use '../../../../../sass/tools' as *;
+
+.orderBody {
+
+}
