+++ src/components/pages/Order/components/OrderNavigation/OrderNavigation.module.scss
@@ -1,0 +1,5 @@
+@use '../../../../../sass/tools' as *;
+
+.orderNavigation {
+
+}
