+++ src/components/pages/Order/components/OrderNavigation/OrderNavigation.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import styles from './OrderNavigation.module.scss';
+
+function OrderNavigation() {
+  return (
+    <div className={styles.orderNavigation}>
+      Навигация
+    </div>
+  );
+}
+
+export default OrderNavigation;
